Batch
Added a large batch of changes. I removed a lot of creatures, which forced me to remove their addon and such to stop errors from occurring. I updated the levels of any creature I could find that wasn't the same values as vMaNGOS is using.

Several of the creatures removed are most likely involved in quests and those need to be reworked.