Stormwind City
Ran through all profession trainers in Stormwind. I might've missed an NPC somewhere, testing needs to be done on these.

The entries in npc_trainer are most likely temporary, they will share ids with other trainers that have the same spells so I'll change that as I go.

I added a text file called Progress.txt that will show what I'm working on and what I have yet to do and what is considered ready for testing.

I added two modules that I use myself. One forces the expansion to be set to classic, max player level to be set 60 and disables the LFG tool. The other module disables achievements. It's not required to use either of these, but I like to be as precise as I can.