Stormwind City
Removed the individual deletes of quest poi entries. Clearing the entire tables instead since the results are the same.

Changed a quest in Stormwind to fail if the player releases their spirit. I tested the quest but it doesn't seem to work properly.

The profession changes also caused some DB errors that I need to look at.