Durotar
Ran through most of Durotar. There are a few creatures left to check, unless I missed some, and the quests needs to be checked.

The western tower in Durotar, which I believe normally has a zeppelin to Thunder Bluff and Borean Tundra, doesn't exist in 1.12 so I cleared the tower and put a gate in the entrance to block it.