Shaman
AoWoW is showing the wrong talents for some reason, a few are removed from the talent tool but they are there in-game. I grabbed the old talents and (for now) just copy-pasted the changed talent on top of the old one with the help of a screenshot from the game.