Mage, Shaman, Warrior
Replaced a talent for Protection Warrior. The revenge talent is useless, as that ability isn't available, so I replaced it with Improved Defensive Stance.
Fixed the image for Shaman talents, nothing important.
Removed Conjure Water Rank 7 from the Mage trainers. That ability is obtained through quests, as far as I know it was indeed added to the trainers in late vanilla but I prefer to not have it there.