Merge
I decided that it was easier to merge everything into one set of folders instead of having them split up. There are a few left that I need to sort still.

I also added the AQ Gate to silithus. I do plan to try my best to port the gate opening event over to azerothcore, but that's a long way down the road.