Module
Merged the different modules into one. It makes it easier for me and right now that's what matters.
Changed the door leading up to UBRS to be closed initially. I need to figure out a way to script the opening with the seal of ascension by use of a module instead of having to modify the core.