Shaman
There is an obvious problem with the final talent of Enhancement Shaman. The talent in 3.3.5 makes both weapons attack, in 1.12 it was "the" weapon. This is because the talent before it is normally dual wield which didn't exist back then, and it means the talent doesn't function at all. I haven't decided what to replace it with but this is my way of showing that the talent is useless until I do decide.