Upper Blackrock Spire
Added the modified script for blackrock spire to require the seal of ascension to enter upper blackrock spire. It's not clean or anywhere near done, but it does open the door only when the player has the item. I need to add the timer to it to trigger the braziers in sequence before the door itself.