The Barrens
Changed the level, rank, faction and loot tables of all creatures (friendly NPCs not included) in the barrens. There are an insane amount of entities in this zone, so checking the respawn times and possibly the spawn points of all creatures and objects will take a very long time.

Added spaces to the folders, it makes no sense to not have them.