Class Spells
Reworked the system I use for the class spells. The new file will delete all spells that require level 61 and above, area of effect buff spells, mounts and the highest available conjure water spell for mage. This one does not work around any talent changes.

I moved the old one to the extra folder, that is the one I'm going to use since it includes my changes to the talents.