Onyxia's Lair
Reworked Onyxia's Lair. The modified script is courtesy of the people developing the official progression module.

This will most likely require some tuning. I also modified the raid on my end to be a 40-man raid and not 10 or 25. That change require dbc changes both on the servers and the clients side.