Progression
Started splitting the files into patches to support progression. The files that are already there have been moved into the first patch folder and I will be changing them over time and move stuff to the proper patch.

Yes, the idea is to move into TBC and WotLK progression eventually but that is extremely far away. After all, most of the zones aren't even done for vanilla yet.